Remove last_edit field from Post model
The last_edit field is redundant in my opinion, if you just
fix most grammer or spelling mistakes the readers don't
have to know when it was edited. If you want to clearify
a point you made in a previous post, create a new post
and add a link to your previous post.

I also made some other changes. I moved setup_server.py to root
and added a seeder.py module to make models.py easier to read
and less cluttered.