Add tabs to markdown editor/preview page
Now it's easier to create or edit posts, the editor
and the HTML preview are now on a seperate tab to
improve readablilty of both. add_post.html and
update_post.html are now merged into one page (post_editor.html).

Also changed this commit:
- JS files now have their own directory.
- Normalize.css has been updated to a new version, there
are way less errors in Opera and Firefox.