Improve import readability and make if statements more clear
I got some warnings that I used tabs and spaces in some files : (.
So I fixed that with the reformat code function, which worked quite
well. Found some weird if statements as well e.g.
if post.visible == True which is useless and doesn't add reall value
to the code in my opinion.