Add propper pagination for the images
Actually I tried to do to much in one commit and fucked it up.
I lost track of some renaming in views/admin.py and had to reset
a bunch of stuff and try to sort the chaos that I created.

So this commit is a bunch of fixes and small changes besides
pagination for the images.