Refactor models.py into a models module with multiple classes
I refactored models.py because it did multiple tasks in one file
that could easily be seperated into multiple classes. It already
contained multiple classes, but now that I'm planning to add some
additional changes it's time to refactor this messy solution.

This also makes it easier to write unit tests for the models.